requirements.txt
# Core orchestration
langgraph>=0.2.28
langchain>=0.2.16
langchain-core>=0.2.38
langchain-anthropic>=0.1.23
langchain-openai>=0.1.25

# Data & schemas
pydantic>=2.7.0
python-dotenv>=1.0.0

# Memory
chromadb>=0.5.5

# CLI
typer>=0.12.5
rich>=13.7.1

# Rendering
pillow>=10.4.0
numpy>=1.26.0

# fal.ai T2V
fal-client>=0.5.0

# Utilities
httpx>=0.27.0

# Web server
fastapi>=0.111.0
uvicorn>=0.30.0
python-multipart>=0.0.9
